Dont pass props when we can connect the component directly
